Bugfixes on threading and constructor
Reordered some of the methods in the constructor which were accessing
attributes not yet initialized. Fixed logic in thread generator which
was simply passing when there was not current blocking thread. Methods
prompting user input have threading disabled for the time being since
when they're run from the interpreter things get wonky.